// WebSite/src/models/models.ts
export class BaseModel {
  id: number = 0;
  creationUserId?: number;
  creationDate?: Date;
  modificationUserId?: number;
  modificationDate?: Date;
}
// ...
export enum StatusType {
  Failed = 0,
  Success = 1,
}
// ...
export class AutoById extends BaseModel {
  id: number = 0;
  name: string = '';
  brandId: number = 0;
  description: string = '';
  model: string = '';
  vinNo: string = '';
  type: number = 0;
  color: string = '';
  engine: string = '';
  lot: string = '';
  carName: string = '';
  buyingAccountId?: number;
  buyAccountName: string = '';
  remainingPayment?: number;
  buyerId?: number = 0;
  buyerName: string = '';
  loadPortId: number = 0;
  loadPortName: string = '';
  auctionId: number = 0;
  auctionName: string = '';
  cityId: number = 0;
  cityName: string = '';
  destinationId: number = 0;
  destinationName: string = '';
  containerId?: number;
  buyDate: Date = new Date();
  buyDateStr: string = '';
  containerArrivalDateStr: string = '';
  carStatus: number = 0;
  paperStatus?: number;
  displayStatus: number = 0;
  isArchive: number = 0;
  carStatusStr: string = '';
  containerSerial: string = '';
  bookNo: string = '';
  departureDateStr: string = '';
  shippingCompany: string = '';
}

export class Payment extends BaseModel {
  clientName: string = '';
  paymentId: number = 0;
  paymentType: number = 0;
  autoId: number = 0;
  vinFormNumber: string = '';
  categoryId: number = 0;
  amount: number = 0;
  payDate?: Date;
  paymentMethod: number = 0;
  notes: string = '';
}
// ...
export class SearchPayment {
  isSearch: boolean = false;
  autoId: number = 0;
  vinNo: string = '';
  purchaseDate!: Date;
  clientId: number = 0;
}